Trailing DOT added to dns_name example
The shown example to create a zone will not work. The api required a trailing dot for the dns_name. In the GCP console this is visible if you display the rest call. However in the documentation it is not. Without the trailing dot the following exception occurs on zone.create():
*** google.api_core.exceptions.BadRequest: 400 POST https://dns.googleapis.com/dns/v1/projects/urlshortener-257711/managedZones: Invalid value for 'entity.managedZone.dnsName': 'test.de'

zone = client.zone('acme-co', 'example.com', <=Missing trailing dot
     ...                    description='Acme Company zone')