docs: add trailing '.' to `dns_name` example (#5)
* docs: add trailing '.' to `dns_name` example

The shown example to create a zone will not work. The api requires a trailing dot for the dns_name. In the GCP console, this is visible if you review the rest call. However, in the documentation it is not. Without the trailing dot the following exception occurs on zone.create():
*** google.api_core.exceptions.BadRequest: 400 POST https://dns.googleapis.com/dns/v1/projects/urlshortener-257711/managedZones: Invalid value for 'entity.managedZone.dnsName': 'example.com'

* docs: Format code in documentation

Co-authored-by: Christopher Wilcox <crwilcox@google.com>